Package structure: moved intervals.
Part of #963.
