+++ src/main/java/de/dennisguse/opentracks/ui/intervals/IntervalStatistics.java
@@ -1,4 +1,4 @@
-package de.dennisguse.opentracks.viewmodels;
+package de.dennisguse.opentracks.ui.intervals;
 
 import androidx.annotation.Nullable;
